コンポーネント整理
