+++ client/src/components/Stock/StockDetail/index.tsx
