+++ client/src/components/Stock/StockList/SearchForm/style.tsx
