+++ client/src/components/Stock/StockList/style.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 
-export const Container = styled.div`
+export const GridContainer = styled.div`
   display: grid;
   grid-template-columns: 120px 120px 120px;
   width: 360px;
